+++ packages/components/src/List/ListComposition/ItemsNumber/ItemsNumber.component.js
@@ -10,13 +10,15 @@ function ItemsNumber(props) {
 			{selected
 				? labelSelected ||
 				  t('LIST_TOOLBAR_NUMBER_OF_SELECTED_ITEMS', {
-						defaultValue: '{{count}}/{{total}} items',
+						defaultValue: '{{count}}/{{total}} item',
+						defaultValue_plural: '{{count}}/{{total}} items',
 						count: selected,
 						total: totalItems,
 				  })
 				: label ||
 				  t('LIST_TOOLBAR_TOTAL_NUMBER_OF_ITEMS', {
-						defaultValue: '{{count}} items',
+						defaultValue: '{{count}} item',
+						defaultValue_plural: '{{count}} items',
 						count: totalItems,
 				  })}
 		</div>