+++ tools/scripts-config-babel/index.js
@@ -1,3 +1,1 @@
-const path = require('path');
-
-module.exports = () => path.join(__dirname, '.babelrc.json');
+module.exports = require('./.babelrc.json');