+++ tools/scripts-config-storybook-lib/.storybook-templates/preview.js
@@ -18,9 +18,13 @@ const i18n = initI18n(userI18n);
 let cmfLoader;
 let cmfDecorator;
 if (cmf) {
-	const cmfPreview = require('./cmf').configureCmfModules(cmf.modules, cmf.settings);
-	cmfLoader = cmfPreview.loader;
-	cmfDecorator = cmfPreview.decorator;
+	try {
+		const cmfPreview = require('./cmf').configureCmfModules(cmf.modules, cmf.settings);
+		cmfLoader = cmfPreview.loader;
+		cmfDecorator = cmfPreview.decorator;
+	} catch(e) {
+		console.error(e);
+	}
 }
 
 const defaultPreview = {