+++ tools/scripts-config-storybook-lib/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @talend/scripts-config-storybook-lib
 
+## 4.0.2
+
+### Patch Changes
+
+- b1cb432e5: fix: support for import of scss file without css module
+
+## 4.0.1
+
+### Patch Changes
+
+- c7152471f: fix: msw is broken
+
+## 4.0.0
+
+### Major Changes
+
+- c18aabb97: chore: simplify it, it just hold the configuration
+
+  BREAKING CHANGE: it do not expose anything from outside.
+
+  This new version do not work with previous talend-scripts version
+
+## 3.0.2
+
+### Patch Changes
+
+- d474e873f: fix: storybook crash trying to load react-cmf module
+
+## 3.0.1
+
+### Patch Changes
+
+- 5e72003aa: fix: make storybook config when there is no cmf package installed
+
+## 3.0.0
+
+### Major Changes
+
+- fe430c316: fix: stories override default stories
+
+  Breaking Change: stories is supposed to give blob to get all the story you need.
+  If the user specify this it is to be able to override it.
+
+  This is a fix to be able to use this configuration to document @talend/design-system package.
+
+### Patch Changes
+
+- 784fe3919: fix: update msw for security reason
+
 ## 2.3.1
 
 ### Patch Changes